chore(project): seperate viewmodel for history view controller
